Added contact panel to exhibitor

+++ exhibitors/views.py
@@ -1,4 +1,6 @@
 from django.forms import modelform_factory
+from django.http import HttpResponseForbidden
+
 from .models import Exhibitor
 from recruitment.models import RecruitmentApplication
 from django.shortcuts import render, redirect, get_object_or_404
@@ -6,17 +8,20 @@ from django.contrib.auth.decorators import permission_required
 
 from orders.models import Order, Product
 
-@permission_required('exhibitors.change_exhibitor', raise_exception=True)
+
 def exhibitors(request, template_name='exhibitors/exhibitors.html'):
 	return render(request, template_name, {'exhibitors': Exhibitor.objects.all().order_by('company__name')})
 
 
-@permission_required('exhibitors.change_exhibitor', raise_exception=True)
+
 def exhibitor(request, pk, template_name='exhibitors/exhibitor.html'):
 	exhibitor = get_object_or_404(Exhibitor, pk=pk)
+	if not request.user.has_perm('exhibitors.change_exhibitor') and request.user not in exhibitor.hosts.all():
+		return HttpResponseForbidden()
+
 	ExhibitorForm = modelform_factory(
 		Exhibitor,
-		exclude=('company', 'fair')
+		exclude=('company', 'fair') if request.user.has_perm('exhibitors.change_exhibitor') else ('company', 'fair', 'hosts', 'contact')
 	)
 	exhibitor_form = ExhibitorForm(request.POST or None, instance=exhibitor)
 	if exhibitor_form.is_valid():
@@ -25,8 +30,8 @@ def exhibitor(request, pk, template_name='exhibitors/exhibitor.html'):
 
 	users = [(recruitment_application.user, recruitment_application.delegated_role)  for recruitment_application in RecruitmentApplication.objects.filter(status='accepted').order_by('user__first_name', 'user__last_name')]
 
-
-	exhibitor_form.fields['hosts'].choices = [('', '---------')] + [(user[0].pk, user[0].get_full_name() + ' - ' + user[1].name) for
+	if request.user.has_perm('exhibitors.change_exhibitor'):
+		exhibitor_form.fields['hosts'].choices = [('', '---------')] + [(user[0].pk, user[0].get_full_name() + ' - ' + user[1].name) for
 																		  user in users]
 
 	return render(request, template_name, {
@@ -35,9 +40,12 @@ def exhibitor(request, pk, template_name='exhibitors/exhibitor.html'):
 	})
 
 
-@permission_required('exhibitors.change_exhibitor', raise_exception=True)
 def order(request, exhibitor_pk, order_pk=None, template_name='exhibitors/order_form.html'):
 	exhibitor = get_object_or_404(Exhibitor, pk=exhibitor_pk)
+
+	if not request.user.has_perm('exhibitors.change_exhibitor') and request.user not in exhibitor.hosts.all():
+		return HttpResponseForbidden()
+
 	OrderFactory = modelform_factory(
 		Order,
 		exclude=('exhibitor',),
@@ -53,8 +61,10 @@ def order(request, exhibitor_pk, order_pk=None, template_name='exhibitors/order_
 	return render(request, template_name, {'form': order_form, 'exhibitor': exhibitor, 'order': order})
 
 
-@permission_required('exhibitors.change_exhibitor', raise_exception=True)
 def order_delete(request, exhibitor_pk, order_pk):
-    order = get_object_or_404(Order, pk=order_pk)
-    order.delete()
-    return redirect('exhibitor', exhibitor_pk)
+	order = get_object_or_404(Order, pk=order_pk)
+	exhibitor = get_object_or_404(Exhibitor, pk=exhibitor_pk)
+	if not request.user.has_perm('exhibitors.change_exhibitor') and request.user not in exhibitor.hosts.all():
+		return HttpResponseForbidden()
+	order.delete()
+	return redirect('exhibitor', exhibitor_pk)
